+++ changelog.md
@@ -6,6 +6,17 @@
 ---
 <br>
 
+### v3.1.0
+### **ANGERY^2**
+---
+* Remove reporting already downloaded file as finished. Instead, redownload it with (1) suffix 
+* Fix the go bindings being unusable
+* Fix directory flattening on Windows
+* Use `byte[]` for the private key in C# bindings
+
+---
+<br>
+
 ### v3.0.0
 ### **ANGERY**
 ---